+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Go server: Add support for `labeled_boolean` metrics with static labels ([AE-1250](https://mozilla-hub.atlassian.net/browse/AE-1250))
+
 ## 19.0.0
 
 - Python server: Support optional metrics ([#828](https://github.com/mozilla/glean_parser/pull/828))